Update hourly-weather-data-test.ipynb
the hour columns only had 0s because that is midnight! had to iterate over all items in the hourly list, not just the first one. Next step is to put date together with hour, in one column. will use datetime in python for that. also, must understand why new df is 840 rows while the rows of the non iterated dataset (34) * hours of the days (24 ) is less than that